% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_accessors.R
\docType{methods}
\name{pixeltype}
\alias{pixeltype}
\alias{pixeltype,antsImage-method}
\alias{pixeltype<-}
\alias{pixeltype<-,antsImage-method}
\alias{components}
\alias{components,antsImage-method}
\alias{components<-}
\alias{components<-,antsImage-method}
\alias{spacing}
\alias{spacing,antsImage-method}
\alias{spacing<-}
\alias{spacing<-,antsImage-method}
\alias{direction}
\alias{direction,antsImage-method}
\alias{direction<-}
\alias{direction<-,antsImage-method}
\alias{origin}
\alias{origin,antsImage-method}
\alias{origin<-}
\alias{origin<-,antsImage-method}
\title{Extract Image Attributes \code{pixeltype} from antsImages}
\usage{
pixeltype(object)

\S4method{pixeltype}{antsImage}(object)

pixeltype(object) <- value

\S4method{pixeltype}{antsImage}(object) <- value

components(object)

\S4method{components}{antsImage}(object)

components(object) <- value

\S4method{components}{antsImage}(object) <- value

spacing(object)

\S4method{spacing}{antsImage}(object)

spacing(object) <- value

\S4method{spacing}{antsImage}(object) <- value

direction(object)

\S4method{direction}{antsImage}(object)

direction(object) <- value

\S4method{direction}{antsImage}(object) <- value

origin(object)

\S4method{origin}{antsImage}(object)

origin(object) <- value

\S4method{origin}{antsImage}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{antsImage}.}

\item{value}{is the value to assign to the slot.}
}
\description{
Methods that act on the \code{slots} of the
antsImage object
}
