% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsAverageImages.R
\name{antsAverageImages}
\alias{antsAverageImages}
\title{Computes average of image list}
\usage{
antsAverageImages(imageList, normalize = FALSE)
}
\arguments{
\item{imageList}{list of antsImages or a character vector of filenames}

\item{normalize}{boolean determines if image is divided by mean before
averaging}
}
\description{
Calculate the mean of a list of antsImages
}
\examples{
f1 = getANTsRData('r16')
f2 = getANTsRData('r64')
r16 <- antsImageRead(f1)
r64 <- antsImageRead(f2)
mylist <- list(r16, r64)
antsAverageImages(mylist)
bad = resampleImage(r16, c(2, 2), useVoxels = FALSE)
antsAverageImages(mylist, normalize = TRUE)
antsAverageImages(c(f1, f2))
}
\author{
Avants BB, Pustina D
}
