% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Arith.R, R/zzz_Comparison.R, R/zzz_Logic.R
\docType{methods}
\name{antsImage-operators}
\alias{antsImage-operators}
\alias{Arith,antsImage,antsImage-method}
\alias{+,antsImage,antsImage-method}
\alias{Arith,antsImage,numeric-method}
\alias{-,antsImage,antsImage-method}
\alias{Arith,antsImage,numeric-method}
\alias{*,antsImage,antsImage-method}
\alias{Arith,antsImage,numeric-method}
\alias{/,antsImage,antsImage-method}
\alias{Arith,antsImage,numeric-method}
\alias{^,antsImage,antsImage-method}
\alias{Arith,antsImage,numeric-method}
\alias{\%\%,antsImage,antsImage-method}
\alias{Arith,antsImage,numeric-method}
\alias{\%/\%,antsImage,antsImage-method}
\alias{Arith,antsImage,numeric-method}
\alias{Arith,antsImage,numeric-method}
\alias{+,antsImage,numeric-method}
\alias{Arith,antsImage,numeric-method}
\alias{-,antsImage,numeric-method}
\alias{Arith,antsImage,numeric-method}
\alias{*,antsImage,numeric-method}
\alias{Arith,antsImage,numeric-method}
\alias{/,antsImage,numeric-method}
\alias{Arith,antsImage,numeric-method}
\alias{^,antsImage,numeric-method}
\alias{Arith,antsImage,numeric-method}
\alias{\%\%,antsImage,numeric-method}
\alias{Arith,antsImage,numeric-method}
\alias{\%/\%,antsImage,numeric-method}
\alias{Arith,antsImage,numeric-method}
\alias{+,numeric,antsImage-method}
\alias{Arith,antsImage,numeric-method}
\alias{-,numeric,antsImage-method}
\alias{Arith,antsImage,numeric-method}
\alias{*,numeric,antsImage-method}
\alias{Arith,antsImage,numeric-method}
\alias{/,numeric,antsImage-method}
\alias{Arith,antsImage,numeric-method}
\alias{^,numeric,antsImage-method}
\alias{Arith,antsImage,numeric-method}
\alias{\%\%,numeric,antsImage-method}
\alias{Arith,antsImage,numeric-method}
\alias{\%/\%,numeric,antsImage-method}
\alias{Arith,antsImage,numeric-method}
\alias{Arith,antsImage,missing-method}
\alias{Arith,antsImage,numeric-method}
\alias{Arith,numeric,antsImage-method}
\alias{Arith,antsImage,logical-method}
\alias{Arith,logical,antsImage-method}
\alias{Arith,antsImage,array-method}
\alias{Arith,array,antsImage-method}
\alias{Arith,list,antsImage-method}
\alias{Arith,antsImage,list-method}
\alias{==,antsImage,antsImage-method}
\alias{==,antsImage,list-method}
\alias{==,antsImage,array-method}
\alias{==,array,antsImage-method}
\alias{==,antsImage,ANY-method}
\alias{==,ANY,antsImage-method}
\alias{>,antsImage,antsImage-method}
\alias{>,antsImage,list-method}
\alias{>,antsImage,array-method}
\alias{>,array,antsImage-method}
\alias{>,antsImage,ANY-method}
\alias{>,ANY,antsImage-method}
\alias{<,antsImage,antsImage-method}
\alias{<,antsImage,list-method}
\alias{<,antsImage,array-method}
\alias{<,array,antsImage-method}
\alias{<,antsImage,ANY-method}
\alias{<,ANY,antsImage-method}
\alias{!=,antsImage,antsImage-method}
\alias{!=,antsImage,list-method}
\alias{!=,antsImage,array-method}
\alias{!=,array,antsImage-method}
\alias{!=,antsImage,ANY-method}
\alias{!=,ANY,antsImage-method}
\alias{<=,antsImage,antsImage-method}
\alias{<=,antsImage,list-method}
\alias{<=,antsImage,array-method}
\alias{<=,array,antsImage-method}
\alias{<=,antsImage,ANY-method}
\alias{<=,ANY,antsImage-method}
\alias{>=,antsImage,antsImage-method}
\alias{>=,antsImage,list-method}
\alias{>=,antsImage,array-method}
\alias{>=,array,antsImage-method}
\alias{>=,antsImage,ANY-method}
\alias{>=,ANY,antsImage-method}
\alias{&,antsImage,antsImage-method}
\alias{Logic,antsImage,antsImage-method}
\alias{|,antsImage,antsImage-method}
\alias{Logic,antsImage,antsImage-method}
\alias{Logic,antsImage,array-method}
\alias{Logic,array,antsImage-method}
\alias{&,antsImage,logical-method}
\alias{Logic,antsImage,logical-method}
\alias{|,antsImage,logical-method}
\alias{Logic,antsImage,logical-method}
\alias{Logic,logical,antsImage-method}
\alias{Logic,antsImage,numeric-method}
\alias{Logic,numeric,antsImage-method}
\alias{Logic,antsImage,ANY-method}
\alias{Logic,ANY,antsImage-method}
\alias{Logic,list,antsImage-method}
\alias{Logic,antsImage,list-method}
\title{Operations for antsImage Objects}
\usage{
\S4method{Arith}{antsImage,antsImage}(e1, e2)

\S4method{+}{antsImage,antsImage}(e1, e2)

\S4method{-}{antsImage,antsImage}(e1, e2)

\S4method{*}{antsImage,antsImage}(e1, e2)

\S4method{/}{antsImage,antsImage}(e1, e2)

\S4method{^}{antsImage,antsImage}(e1, e2)

\S4method{\%\%}{antsImage,antsImage}(e1, e2)

\S4method{\%/\%}{antsImage,antsImage}(e1, e2)

\S4method{Arith}{antsImage,numeric}(e1, e2)

\S4method{+}{antsImage,numeric}(e1, e2)

\S4method{-}{antsImage,numeric}(e1, e2)

\S4method{*}{antsImage,numeric}(e1, e2)

\S4method{/}{antsImage,numeric}(e1, e2)

\S4method{^}{antsImage,numeric}(e1, e2)

\S4method{\%\%}{antsImage,numeric}(e1, e2)

\S4method{\%/\%}{antsImage,numeric}(e1, e2)

\S4method{+}{numeric,antsImage}(e1, e2)

\S4method{-}{numeric,antsImage}(e1, e2)

\S4method{*}{numeric,antsImage}(e1, e2)

\S4method{/}{numeric,antsImage}(e1, e2)

\S4method{^}{numeric,antsImage}(e1, e2)

\S4method{\%\%}{numeric,antsImage}(e1, e2)

\S4method{\%/\%}{numeric,antsImage}(e1, e2)

\S4method{Arith}{antsImage,missing}(e1, e2)

\S4method{Arith}{numeric,antsImage}(e1, e2)

\S4method{Arith}{antsImage,logical}(e1, e2)

\S4method{Arith}{logical,antsImage}(e1, e2)

\S4method{Arith}{antsImage,array}(e1, e2)

\S4method{Arith}{array,antsImage}(e1, e2)

\S4method{Arith}{list,antsImage}(e1, e2)

\S4method{Arith}{antsImage,list}(e1, e2)

\S4method{==}{antsImage,antsImage}(e1, e2)

\S4method{==}{antsImage,list}(e1, e2)

\S4method{==}{antsImage,array}(e1, e2)

\S4method{==}{array,antsImage}(e1, e2)

\S4method{==}{antsImage,ANY}(e1, e2)

\S4method{==}{ANY,antsImage}(e1, e2)

\S4method{>}{antsImage,antsImage}(e1, e2)

\S4method{>}{antsImage,list}(e1, e2)

\S4method{>}{antsImage,array}(e1, e2)

\S4method{>}{array,antsImage}(e1, e2)

\S4method{>}{antsImage,ANY}(e1, e2)

\S4method{>}{ANY,antsImage}(e1, e2)

\S4method{<}{antsImage,antsImage}(e1, e2)

\S4method{<}{antsImage,list}(e1, e2)

\S4method{<}{antsImage,array}(e1, e2)

\S4method{<}{array,antsImage}(e1, e2)

\S4method{<}{antsImage,ANY}(e1, e2)

\S4method{<}{ANY,antsImage}(e1, e2)

\S4method{!=}{antsImage,antsImage}(e1, e2)

\S4method{!=}{antsImage,list}(e1, e2)

\S4method{!=}{antsImage,array}(e1, e2)

\S4method{!=}{array,antsImage}(e1, e2)

\S4method{!=}{antsImage,ANY}(e1, e2)

\S4method{!=}{ANY,antsImage}(e1, e2)

\S4method{<=}{antsImage,antsImage}(e1, e2)

\S4method{<=}{antsImage,list}(e1, e2)

\S4method{<=}{antsImage,array}(e1, e2)

\S4method{<=}{array,antsImage}(e1, e2)

\S4method{<=}{antsImage,ANY}(e1, e2)

\S4method{<=}{ANY,antsImage}(e1, e2)

\S4method{>=}{antsImage,antsImage}(e1, e2)

\S4method{>=}{antsImage,list}(e1, e2)

\S4method{>=}{antsImage,array}(e1, e2)

\S4method{>=}{array,antsImage}(e1, e2)

\S4method{>=}{antsImage,ANY}(e1, e2)

\S4method{>=}{ANY,antsImage}(e1, e2)

\S4method{&}{antsImage,antsImage}(e1, e2)

\S4method{|}{antsImage,antsImage}(e1, e2)

\S4method{Logic}{antsImage,array}(e1, e2)

\S4method{Logic}{array,antsImage}(e1, e2)

\S4method{&}{antsImage,logical}(e1, e2)

\S4method{|}{antsImage,logical}(e1, e2)

\S4method{Logic}{logical,antsImage}(e1, e2)

\S4method{Logic}{antsImage,numeric}(e1, e2)

\S4method{Logic}{numeric,antsImage}(e1, e2)

\S4method{Logic}{antsImage,ANY}(e1, e2)

\S4method{Logic}{ANY,antsImage}(e1, e2)

\S4method{Logic}{list,antsImage}(e1, e2)

\S4method{Logic}{antsImage,list}(e1, e2)
}
\arguments{
\item{e1}{is an object of class \code{antsImage}.}

\item{e2}{is an object of class \code{antsImage}.}
}
\description{
Overloaded operators for antsImage objects
}
\examples{

img01 <- as.antsImage(array(1:64, c(4,4,4,1)))
img02 <- as.antsImage(array(64:1, c(4,4,4,1)))
is.antsImage(img01 + img02)
is.antsImage(sqrt(2) * img01)
is.antsImage(img02 / pi)
}
\author{
John Muschellli \email{muschellij2@gmail.com}
}
