% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Math_Summary.R
\docType{methods}
\name{max}
\alias{Summary,antsImage-method}
\alias{all}
\alias{antsImage-summary}
\alias{any}
\alias{max}
\alias{min}
\alias{prod}
\alias{range}
\alias{sum}
\title{Summary for antsImage Objects}
\usage{
max(x, ..., na.rm = FALSE)

min(x, ..., na.rm = FALSE)

range(x, ..., na.rm = FALSE)

prod(x, ..., na.rm = FALSE)

sum(x, ..., na.rm = FALSE)

any(x, ..., na.rm = FALSE)

all(x, ..., na.rm = FALSE)

\S4method{Summary}{antsImage}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{is an object of class \code{antsImage}.}

\item{...}{further arguments passed to summary methods}

\item{na.rm}{logical: should missing values be removed?}
}
\description{
Overloaded Summary for antsImage objects
}
\examples{
img01 <- as.antsImage(array(1:64, c(4,4,4,1)))
max(img01)
min(img01)
range(img01)
prod(img01)
}

