% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrImageToImageMetric_class.R
\name{antsrImageToImageMetric.GetDerivative}
\alias{antsrImageToImageMetric.GetDerivative}
\title{antsrImageToImageMetric.GetDerivative}
\usage{
antsrImageToImageMetric.GetDerivative(metric)
}
\arguments{
\item{metric}{an 'antsrImageToImageMetric'}
}
\value{
image similarity value
}
\description{
get derivative of image metric
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrImageToImageMetric.Create(x,y,type="MeanSquares")
tx <- createAntsrTransform( precision="double", type="AffineTransform", dimension=2)
setAntsrTransformParameters(tx, c(0,-1,1,0,0,0))
setAntsrTransformFixedParameters(tx, c(128,128))
antsrImageToImageMetric.SetMovingTransform(metric, tx)
metricValue = antsrImageToImageMetric.GetDerivative(metric)
}
