% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrImageToImageMetric_class.R
\name{antsrImageToImageMetric.Initialize}
\alias{antsrImageToImageMetric.Initialize}
\title{antsrImageToImageMetric.Initialize}
\usage{
antsrImageToImageMetric.Initialize(metric)
}
\arguments{
\item{metric}{an 'antsrImageToImageMetric'}
}
\description{
set moving image mask for image to image metric
}
\note{
must call this after setting up object, before getting values back
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
z = getMask(y)
antsrImageToImageMetric.SetMovingImageMask(metric, z)
antsrImageToImageMetric.Initialize(metric)
}

