% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrImageToImageMetric_class.R
\name{antsrImageToImageMetric.SetFixedImage}
\alias{antsrImageToImageMetric.SetFixedImage}
\title{antsrImageToImageMetric.SetFixedImage}
\usage{
antsrImageToImageMetric.SetFixedImage(metric, image)
}
\arguments{
\item{metric}{an 'antsrImageToImageMetric'}

\item{image}{the fixed 'antsImage'}
}
\description{
set fixed image for image to image metric
}
\note{
After calling this, must call antsrImageToImageMetric.Initialize(metric)
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrImageToImageMetric.Create(x,y,type="MeanSquares")
z = x*2
antsrImageToImageMetric.SetFixedImage(metric, z)
}

