% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrImageToImageMetric_class.R
\name{antsrImageToImageMetric.SetFixedImageMask}
\alias{antsrImageToImageMetric.SetFixedImageMask}
\title{antsrImageToImageMetric.SetFixedImageMask}
\usage{
antsrImageToImageMetric.SetFixedImageMask(metric, mask)
}
\arguments{
\item{metric}{an 'antsrImageToImageMetric'}

\item{image}{the fixed mask 'antsImage'}
}
\description{
set fixed image mask for image to image metric
}
\note{
After calling this, must call antsrImageToImageMetric.Initialize(metric)
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrImageToImageMetric.Create(x,y,type="MeanSquares")
z = getMask(x)
antsrImageToImageMetric.SetFixedImageMask(metric, z)
}

