% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrImageToImageMetric_class.R
\name{antsrImageToImageMetric.SetMovingImage}
\alias{antsrImageToImageMetric.SetMovingImage}
\title{antsrImageToImageMetric.SetMovingImage}
\usage{
antsrImageToImageMetric.SetMovingImage(metric, image, isMask = FALSE)
}
\arguments{
\item{metric}{an 'antsrImageToImageMetric'}

\item{image}{the moving 'antsImage'}

\item{isMask}{flag indicating if input is image or mask}
}
\description{
set moving image for image to image metric
}
\note{
After calling this, must call antsrImageToImageMetric.Initialize(metric)
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrImageToImageMetric.Create(x,y,type="MeanSquares")
z = y*2
antsrImageToImageMetric.SetMovingImage(metric, z)
}
