% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrImageToImageMetric_class.R
\name{antsrImageToImageMetric.SetMovingImageMask}
\alias{antsrImageToImageMetric.SetMovingImageMask}
\title{antsrImageToImageMetric.SetMovingImageMask}
\usage{
antsrImageToImageMetric.SetMovingImageMask(metric, image)
}
\arguments{
\item{metric}{an 'antsrImageToImageMetric'}

\item{image}{the moving 'antsImage'}
}
\description{
set moving image for image to image metric
}
\note{
After calling this, must call antsrImageToImageMetric.Initialize(metric)
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrImageToImageMetric.Create(x,y,type="MeanSquares")
z = getMask(y)
antsrImageToImageMetric.SetMovingImageMask(metric, z)
}
