% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrImageToImageMetric_class.R
\name{antsrImageToImageMetric.SetMovingTransform}
\alias{antsrImageToImageMetric.SetMovingTransform}
\title{antsrImageToImageMetric.SetMovingTransform}
\usage{
antsrImageToImageMetric.SetMovingTransform(metric, transform)
}
\arguments{
\item{metric}{an 'antsrImageToImageMetric'}

\item{transform}{an 'antsrTransform'}
}
\description{
set transform for moving image
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrImageToImageMetric.Create(x,y,type="ANTSNeighborhoodCorrelation")
tx <- createAntsrTransform( precision="double", type="AffineTransform", dimension=2)
setAntsrTransformParameters(tx, c(0,-1,1,0,0,0))
setAntsrTransformFixedParameters(tx, c(128,128))
antsrImageToImageMetric.SetMovingTransform(metric, tx)
antsrImageToImageMetric.GetValue(metric)
}
