% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrImageToImageMetric_class.R
\name{antsrImageToImageMetric.SetNumberOfHistogramBins}
\alias{antsrImageToImageMetric.SetNumberOfHistogramBins}
\title{antsrImageToImageMetric.SetNumberOfHistogramBins}
\usage{
antsrImageToImageMetric.SetNumberOfHistogramBins(metric, nBins)
}
\arguments{
\item{metric}{an 'antsrImageToImageMetric' of type 'MattesMutualInformation' or 'JointHistogramMutualInformation'}

\item{nBins}{number of bins (minimum is 5 even for binary data)}
}
\description{
set histogram bins image to image metric
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrImageToImageMetric.Create(x,y,type="MattesMutualInformation")
antsrImageToImageMetric.SetNumberOfHistogramBins(metric,12)
}
