% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrImageToImageMetric_class.R
\name{antsrImageToImageMetric.SetRadius}
\alias{antsrImageToImageMetric.SetRadius}
\title{antsrImageToImageMetric.SetRadius}
\usage{
antsrImageToImageMetric.SetRadius(metric, radius)
}
\arguments{
\item{metric}{an 'antsrImageToImageMetric' of type 'ANTSNeighborhoodCorrelation'}

\item{radius}{radius of neighborhood}
}
\description{
set neighborhood radius
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrImageToImageMetric.Create(x,y,type="ANTSNeighborhoodCorrelation")
antsrImageToImageMetric.SetRadius(metric,5)
}
