% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMetric_class.R
\name{antsrMetric.GetDerivative}
\alias{antsrMetric.GetDerivative}
\title{antsrMetric.GetDerivative}
\usage{
antsrMetric.GetDerivative(metric)
}
\arguments{
\item{metric}{an 'antsrMetric'}
}
\value{
image similarity value
}
\description{
get derivative of image metric
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrMetric.Create(x,y,type="MeanSquares")
tx <- createAntsrTransform( precision="double", type="AffineTransform", dimension=2)
setAntsrTransformParameters(tx, c(0,-1,1,0,0,0))
setAntsrTransformFixedParameters(tx, c(128,128))
antsrMetric.SetMovingTransform(metric, tx)
metricValue = antsrMetric.GetDerivative(metric)
}
