% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMetric_class.R
\name{antsrMetric.GetValue}
\alias{antsrMetric.GetValue}
\title{antsrMetric.GetValue}
\usage{
antsrMetric.GetValue(metric)
}
\arguments{
\item{metric}{an 'antsrMetric'}
}
\value{
image similarity value
}
\description{
get current value of metric
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrMetric.Create(x,y,type="MeanSquares")
metricValue = antsrMetric.GetValue(metric)
}
