% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMetric_class.R
\name{antsrMetric.Initialize}
\alias{antsrMetric.Initialize}
\title{antsrMetric.Initialize}
\usage{
antsrMetric.Initialize(metric)
}
\arguments{
\item{metric}{an 'antsrMetric'}
}
\description{
prepare to return values
}
\note{
must call this after setting up object, before getting values back
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
z = getMask(y)
metric = antsrMetric.Create(x,y,type="MeanSquares")
antsrMetric.SetMovingImageMask(metric, z)
antsrMetric.Initialize(metric)
}
