% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMetric_class.R
\name{antsrMetric.SetFixedImageMask}
\alias{antsrMetric.SetFixedImageMask}
\title{antsrMetric.SetFixedImageMask}
\usage{
antsrMetric.SetFixedImageMask(metric, image)
}
\arguments{
\item{metric}{an 'antsrMetric'}

\item{image}{the fixed 'antsImage'}
}
\description{
set fixed image for image to image metric
}
\note{
After calling this, must call antsrMetric.Initialize(metric)
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrMetric.Create(x,y,type="MeanSquares")
z = getMask(x)
antsrMetric.SetFixedImageMask(metric, z)
}
