% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMetric_class.R
\name{antsrMetric.SetMovingImage}
\alias{antsrMetric.SetMovingImage}
\title{antsrMetric.SetMovingImage}
\usage{
antsrMetric.SetMovingImage(metric, image)
}
\arguments{
\item{metric}{an 'antsrMetric'}

\item{image}{the moving 'antsImage'}
}
\description{
set moving image for image to image metric
}
\note{
After calling this, must call antsrMetric.Initialize(metric)
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrMetric.Create(x,y,type="MeanSquares")
z = y*2
antsrMetric.SetMovingImage(metric, z)
}
