% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMetric_class.R
\name{antsrMetric.SetMovingTransform}
\alias{antsrMetric.SetMovingTransform}
\title{antsrMetric.SetMovingTransform}
\usage{
antsrMetric.SetMovingTransform(metric, transform)
}
\arguments{
\item{metric}{an 'antsrMetric'}

\item{transform}{an 'antsrTransform'}
}
\description{
set transform for moving image
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrMetric.Create(x,y,type="ANTSNeighborhoodCorrelation")
tx <- createAntsrTransform( precision="double", type="AffineTransform", dimension=2)
setAntsrTransformParameters(tx, c(0,-1,1,0,0,0))
setAntsrTransformFixedParameters(tx, c(128,128))
antsrMetric.SetMovingTransform(metric, tx)
antsrMetric.GetValue(metric)
}
