% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMetric_class.R
\name{antsrMetric.SetNumberOfHistogramBins}
\alias{antsrMetric.SetNumberOfHistogramBins}
\title{antsrMetric.SetNumberOfHistogramBins}
\usage{
antsrMetric.SetNumberOfHistogramBins(metric, nBins)
}
\arguments{
\item{metric}{an 'antsrMetric' of type 'MattesMutualInformation' or 'JointHistogramMutualInformation'}

\item{nBins}{number of bins (minimum is 5 even for binary data)}
}
\description{
set histogram bins image to image metric
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrMetric.Create(x,y,type="MattesMutualInformation")
antsrMetric.SetNumberOfHistogramBins(metric,12)
}
