% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMetric_class.R
\name{antsrMetric.SetRadius}
\alias{antsrMetric.SetRadius}
\title{antsrMetric.SetRadius}
\usage{
antsrMetric.SetRadius(metric, radius)
}
\arguments{
\item{metric}{an 'antsrMetric' of type 'ANTSNeighborhoodCorrelation'}

\item{radius}{radius of neighborhood}
}
\description{
set neighborhood radius
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrMetric.Create(x,y,type="ANTSNeighborhoodCorrelation")
antsrMetric.SetRadius(metric,5)
}
