% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMetric_class.R
\name{antsrMetric.SetSampling}
\alias{antsrMetric.SetSampling}
\title{antsrMetric.SetSampling}
\usage{
antsrMetric.SetSampling(metric, sampling.strategy, sampling.percentage)
}
\arguments{
\item{metric}{an 'antsrMetric'}

\item{sampling.strategy}{sampling strategy, default if full sampling
\itemize{
  \item{none}{Full sampling}
  \item{random}{}
  \item{regular}{}
}}

\item{sampling.percentage}{percentage of data to sample when calculating metric}
}
\description{
set image sampling strategy and rate
}
\note{
After calling this, must call antsrMetric.Initialize(metric)
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrMetric.Create(x,y,type="MeanSquares")
antsrMetric.SetSampling(metric,"random",0.4)
}
