% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrTransform_class.R
\name{displacementFieldFromAntsrTransform}
\alias{displacementFieldFromAntsrTransform}
\title{displacementFieldFromAntsrTransform}
\usage{
displacementFieldFromAntsrTransform(tx)
}
\arguments{
\item{tx}{antsrTransform}
}
\value{
antsImage
}
\description{
Conver antsrTransform to displacement field
}
