% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigLMStats.R
\name{ilr.predict}
\alias{ilr.predict}
\title{Predict from ilr output}
\usage{
ilr.predict(ilrResult, dataFrame, voxmats, myFormula)
}
\arguments{
\item{ilrResult}{This output form ilr}

\item{dataFrame}{This data frame contains all relevant predictors except for
the matrices associated with the image variables.}

\item{voxmats}{The named list of matrices that contains the changing predictors.}

\item{myFormula}{This is a character string that defines a valid regression formula.}
}
\value{
the predicted matrix.
}
\description{
This function computes a prediction given \code{ilr} output.
}
\examples{

set.seed(1500)
nsub = 100
outcome = rnorm( nsub )
covar = rnorm( nsub )
mat = replicate( nsub, rnorm( nsub ) )
mat2 = replicate( nsub, rnorm( nsub ) )
myform = " outcome ~ covar + vox "
df = data.frame( outcome = outcome, covar = covar )
result = ilr( df, list( vox = mat ), myform)
print( names( result ) )
print( rownames( result$pValue ) )
myform = " vox2 ~ covar * vox "
df = data.frame( outcome = outcome, covar = covar )
result = ilr( df, list( vox = mat, vox2=mat2 ), myform)
df = data.frame(  covar = covar )
pred = ilr.predict(  result, df, list( vox = mat, vox2=mat2 ), myform )

}
\seealso{
\code{\link{ilr}}
}
\author{
BB Avants.
}
