% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_antsProcessArguments.R
\name{.int_antsProcessArguments}
\alias{.int_antsProcessArguments}
\title{args = "hey"
.int_antsProcessArguments(args)
args = list(h = "hey")
.int_antsProcessArguments(args)
args = list(h = list("hey"))
.int_antsProcessArguments(args)}
\usage{
.int_antsProcessArguments(args)
}
\description{
args = "hey"
.int_antsProcessArguments(args)
args = list(h = "hey")
.int_antsProcessArguments(args)
args = list(h = list("hey"))
.int_antsProcessArguments(args)
}
