% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdImage.R
\name{integrateVelocityField}
\alias{integrateVelocityField}
\title{Integrate velocity field}
\usage{
integrateVelocityField(
  referenceImage,
  velocityFieldFileName,
  deformationFieldFileName,
  lowerTime = 0,
  upperTime = 1,
  deltaTime = 0.01
)
}
\arguments{
\item{referenceImage}{defines the image domain}

\item{velocityFieldFileName}{the velocity field exists on disk.}

\item{deformationFieldFileName}{the deformation field output file name.}

\item{lowerTime}{the starting time, usually zero for forward transformation
and one for the inverse transformation.}

\item{upperTime}{the ending time, usually one for forward transformation
and zero for the inverse transformation.}

\item{deltaTime}{the integration time step}
}
\description{
Utility function to integrate a velocity field and create a deformation field.
}
\author{
Avants BB
}
