% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{labelStatsR}
\alias{labelStatsR}
\title{Rcpp interface to labelStatsR}
\usage{
labelStatsR(r_image, r_labelImage)
}
\arguments{
\item{r_image}{S4 antsImage object used as the input image.}

\item{r_labelImage}{S4 antsImage label image defining regions of interest.}
}
\value{
Return value from the underlying C++ routine.
}
\description{
Low-level Rcpp entry point for the ANTsRCore C++ routine \code{labelStatsR}.
}
\details{
This function is generated by Rcpp and is exported for compatibility with
ANTsR and other low-level callers. Most users should call the higher-level
ANTsR wrapper that validates arguments and documents the supported workflow.
}
\keyword{internal}
