% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Operators.R
\name{median.antsImage}
\alias{median.antsImage}
\title{Median for antsImage Objects}
\usage{
\method{median}{antsImage}(x, na.rm = FALSE, mask)
}
\arguments{
\item{x}{is an object of class \code{antsImage}.}

\item{na.rm}{a logical value indicating whether NA should be removed}

\item{mask}{binary mask of values to subset}
}
\description{
Overloaded Median for antsImage objects
}
