% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n3BiasFieldCorrection.R
\name{n3BiasFieldCorrection2}
\alias{n3BiasFieldCorrection2}
\title{N3 Bias field correction}
\usage{
n3BiasFieldCorrection2(
  img,
  mask,
  shrinkFactor = 4,
  convergence = list(iters = 50, tol = 0.0000001),
  splineParam = 200,
  numberOfFittingLevels = 4,
  returnBiasField = FALSE,
  verbose = FALSE,
  weight_mask = NULL
)
}
\arguments{
\item{img}{input antsImage}

\item{mask}{input mask, if one is not passed one will be made}

\item{shrinkFactor}{Shrink factor for multi-resolution correction,
typically integer less than 4}

\item{convergence}{List of:  \code{iters}, maximum number of
iterations and \code{tol}, the convergence tolerance.}

\item{splineParam}{Parameter controlling number of control points in spline.
Either single value, indicating how many control points, or vector
with one entry per dimension of image, indicating the spacing in each direction.}

\item{returnBiasField}{bool, return the field instead of the corrected image.}

\item{verbose}{enables verbose output.}

\item{weight_mask}{antsImage of weight mask}
}
\value{
bias corrected image or bias field
}
\description{
Perform N3 bias field correction on the given image
}
\examples{
 dims = c(50, 50)
 img<-makeImage(imagesize = dims, rnorm(prod(dims)) )
 n3img<-n3BiasFieldCorrection2(img)
 n3img<-n3BiasFieldCorrection2(img, mask = img > 0)
 testthat::expect_error(n3BiasFieldCorrection2(img, weight_mask = "somepath"))
 testthat::expect_error(n3BiasFieldCorrection2(img, splineParam = rep(200, 3)))
 n3img<-n3BiasFieldCorrection2(img, splineParam = c(200, 20))

 rm(img); gc()
 rm(n3img); gc()
fname = getANTsRData("r16")
in_img = antsImageRead(fname)
n3 = n3BiasFieldCorrection2(in_img)
rm(n3); gc()
mask = in_img > 0
mask2 = antsImageClone(mask, out_pixeltype = "float")
# fails
mask
sum(mask)
n3 = n3BiasFieldCorrection2(in_img, mask = mask, verbose = TRUE)
# fails
n3 = n3BiasFieldCorrection2(in_img, mask = mask2)
}
\author{
Avants BB, Tustison NJ
}
