% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripmmarc.R
\name{ripmmarcBasisImage}
\alias{ripmmarcBasisImage}
\title{Generate an antsImage from ripmmarc basis data}
\usage{
ripmmarcBasisImage(canonicalFrame, patchBasisVector, eps = 1e-12)
}
\arguments{
\item{canonicalFrame}{canonical frame image generated by ripmmarc}

\item{patchBasisVector}{the basis vector to convert to image space}

\item{eps}{small value threshold to help generate mask from canonicalFrame}
}
\value{
antsImage
}
\description{
This function converts a vectorized image patch back into an
antsImage such that it may be displayed in its extrinsic dimensionality.
}
\examples{

img <- antsImageRead( getANTsRData( "rand" ) )
img = resampleImage( img, c( 32, 32 ) )
msk <- thresholdImage( img, quantile( img[ img > 0 ], 0.5 )[1], max( img ) )
ripped <- ripmmarc( img, msk, patchRadius=1, patchSamples=20, patchVarEx = 2 )
bimg = ripmmarcBasisImage( ripped$canonicalFrame, ripped$basisMat[1,] )

}
\author{
Kandel BM, Avants BB
}
