% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripmmarc.R
\name{scaleSpaceFeatureDetection}
\alias{scaleSpaceFeatureDetection}
\title{Scale space feature detector}
\usage{
scaleSpaceFeatureDetection(
  image,
  numberOfBlobsToExtract,
  minScale = 1,
  maxScale = 2^10,
  stepsPerOctave = 10
)
}
\arguments{
\item{image}{the input image}

\item{numberOfBlobsToExtract}{the estimated blob count}

\item{minScale}{the minimum amount of smoothing in scale space}

\item{maxScale}{the maximum amount of smoothing in scale space}

\item{stepsPerOctave}{the number of steps to take over scale space octaves}
}
\value{
list of antsImage and dataframe for blob descriptors
}
\description{
Deploy a multiscale laplacian blob detector on an image.  The
function will return the blob descriptors as an image and data frame.
}
\examples{

# WIP: should explore effect of parameters further
blob1 = scaleSpaceFeatureDetection( ri( 1 ), 50 )
blob2 = scaleSpaceFeatureDetection( max( ri( 1 ) ) - ri( 1 ), 50 )

}
\author{
Avants BB
}
