% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Summary.R
\name{sd}
\alias{sd}
\alias{sd.default}
\alias{sd.antsImage}
\title{SD generic}
\usage{
sd(x, ...)

\method{sd}{default}(x, ...)

\method{sd}{antsImage}(x, mask = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{an object for which we want to compute the SD}

\item{\dots}{Any additional arguments to be passed to \code{sd}.}

\item{mask}{is an object of class \code{antsImage}}

\item{na.rm}{a logical value indicating whether NA should be removed}

\item{x}{is an object of class \code{antsImage}.}
}
\description{
Calculates the SD of an image

Overloaded SD for antsImage objects
}
\examples{
img <- antsImageRead( getANTsRData( "r16" ) )
sd(img)
}
