% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Math_Summary.R
\name{var}
\alias{var}
\alias{var.antsImage}
\alias{var.default}
\title{Variance generic}
\usage{
var(x, ...)

\method{var}{default}(x, ...)

\method{var}{antsImage}(x, ...)
}
\arguments{
\item{x}{an object for which we want to compute the variance}

\item{\dots}{Any additional arguments to be passed to \code{var}.}
}
\description{
Calculates the variance of an image
}

