% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrImageToImageMetric_class.R
\name{antsrImageToImageMetric.GetValue}
\alias{antsrImageToImageMetric.GetValue}
\title{antsrImageToImageMetric.GetValue}
\usage{
antsrImageToImageMetric.GetValue(metric)
}
\arguments{
\item{metric}{an 'antsrImageToImageMetric'}
}
\value{
image similarity value
}
\description{
get current value of metric
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrImageToImageMetric.Create(x,y,type="MeanSquares")
metricValue = antsrImageToImageMetric.GetValue(metric)
}
