% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoiseImage.R
\name{denoiseImage}
\alias{denoiseImage}
\title{Denoise an image}
\usage{
denoiseImage(img, mask, shrinkFactor = 1, p = 1, r = 3,
  noiseModel = c("Rician", "Gaussian"))
}
\arguments{
\item{img}{scalar image to denoise.}

\item{mask}{to limit the denoise region.}

\item{shrinkFactor}{downsampling level performed within the algorithm.}

\item{p}{patch radius for local sample.}

\item{r}{search radius from which to choose extra local samples.}

\item{noiseModel}{either Rician or Gaussian.}
}
\value{
antsImage denoised version of image
}
\description{
Denoise an image using a spatially adaptive filter originally described in J. V.
Manjon, P. Coupe, Luis Marti-Bonmati, D. L. Collins, and M. Robles. Adaptive
Non-Local Means Denoising of MR Images With Spatially Varying Noise Levels,
Journal of Magnetic Resonance Imaging, 31:192-203, June 2010.
}
\examples{

img <- antsImageRead( getANTsRData("rand")  ) \%>\% resampleImage( c(32, 32) )
dimg <- denoiseImage( img, img * 0 + 1 )

}
\author{
N Tustison, B Avants
}
