% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigLMStats.R
\name{ilr.predict}
\alias{ilr.predict}
\title{Predict from ilr output}
\usage{
ilr.predict(ilrResult, dataFrame, voxmats, myFormula)
}
\arguments{
\item{ilrResult}{This output form ilr}

\item{dataFrame}{This data frame contains all relevant predictors except for
the matrices associated with the image variables.}

\item{voxmats}{The named list of matrices that contains the changing predictors.}

\item{myFormula}{This is a character string that defines a valid regression formula.}
}
\value{
the predicted matrix.
}
\description{
This function computes a prediction given \code{ilr} output.
}
\examples{

set.seed(1500)
nsub = 100
covar = rnorm( nsub )
mat = replicate( nsub, rnorm( nsub ) )
mat2 = replicate( nsub, rnorm( nsub ) )
outcome = mat[ , 2 ]
myform = " outcome ~ covar + vox + vox2 "
df = data.frame( outcome = outcome, covar = covar )
result = ilr( df, list( vox = mat, vox2 = mat2 ), myform)
df2 = data.frame(  covar = covar )
pred = ilr.predict(  result, df2, list( vox = mat, vox2=mat2 ), myform )
max( cor( result$predictions, df$outcome ) )
max( cor( pred, df$outcome ) )
tail( diag( cor( result$predictions, pred ) ) )

}
\seealso{
\code{\link{ilr}}
}
\author{
BB Avants.
}
