% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointLabelFusion.R
\name{localJointLabelFusion}
\alias{localJointLabelFusion}
\title{local joint label and intensity fusion}
\usage{
localJointLabelFusion(
  targetI,
  whichLabel,
  template,
  templateLabels,
  atlasList,
  labelList,
  submaskDilation = 10,
  typeofTransform = "SyN",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{targetI}{antsImage to be approximated}

\item{whichLabel}{a label number that exists in both the template and library}

\item{template}{a label number that exists in both the template and library}

\item{templateLabels}{a label number that exists in both the template and library}

\item{atlasList}{list containing antsImages with intensity images}

\item{labelList}{optional list containing antsImages with segmentation labels}

\item{typeofTransform}{passed to \code{antsRegistration}.}

\item{verbose}{boolean}

\item{...}{extra parameters passed to JLF}
}
\value{
label probabilities
}
\description{
A local version of joint label fusion that focuses on a specific label.
This is primarily different from standard JLF because it performs registration
on a per label basis and focuses JLF on that label alone.
}
\examples{

}
\author{
Brian B. Avants
}
\keyword{fusion,}
\keyword{template}
