% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Summary.R
\name{unique.antsImage}
\alias{unique.antsImage}
\title{Unique for antsImage Objects}
\usage{
\method{unique}{antsImage}(x, incomparables = FALSE, ..., mask)
}
\arguments{
\item{x}{is an object of class \code{antsImage}.}

\item{incomparables}{a vector of values that cannot be compared.}

\item{...}{additional arguments passed to \code{\link{unique}}}

\item{mask}{binary mask of values to subset}
}
\description{
Overloaded uniqueness operation for antsImage objects
}
\examples{
img <- antsImageRead( getANTsRData( "r16" ) )
img[img > 5] = 0
sort(unique(img))

}
